chore: make ubergen a library
Move the useful stuff in ubergen into lib/index.ts and add `main` field
to package json so packages within the repository can depend on it's
functionality. Specifically useful for repo restructure work.
